+++ sidebars.js
@@ -182,6 +182,11 @@ const sidebars = {
       ],
     },
   ],
+  gitaliasSidebar: [
+    { type: 'doc', id: 'gitalias/index', label: "Intro" },
+    { type: 'doc', id: 'gitalias/commands-aliases', label: "Commands Aliases" },
+    { type: 'doc', id: 'gitalias/most-used-aliases', label: "Most Used Aliases" },
+  ],
 };
 
 module.exports = sidebars;