+++ CHANGELOG.md
@@ -7,10 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add documentation for gitalias ([#4](https://github.com/josefpihrt/josefpihrt.github.io/pull/4)).
+
 ## [0.1.0] - 2023-07-23
 
 ### Added
 
-- Add GitHub icon to navbar
-- Make sidebar hideable
-- Add documentation for [Roslynator](https://github.com/josefpihrt/roslynator), [Orang](https://github.com/josefpihrt/orang) and [DotMarkdown](https://github.com/josefpihrt/dotmarkdown)
+- Add GitHub icon to navbar ([#3](https://github.com/josefpihrt/josefpihrt.github.io/pull/3)).
+- Make sidebar hideable ([#3](https://github.com/josefpihrt/josefpihrt.github.io/pull/3)).
+- Add documentation for [Roslynator](https://github.com/josefpihrt/roslynator), [Orang](https://github.com/josefpihrt/orang) and [DotMarkdown](https://github.com/josefpihrt/dotmarkdown) ([#1](https://github.com/josefpihrt/josefpihrt.github.io/pull/1)).